requirements.txt
python-telegram-bot>=20.0
pillow>=9.0.0
qrcode>=7.3.1
python-dotenv>=0.20.0